pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-os"
version = "1.0.0"
description = "Autonomous HRM agent with Ubuntu OS integration"
authors = [{name = "Wulfic"}]
requires-python = ">=3.10"
license = { text = "AI-OS Non‑Selling Attribution License (ANSL) v1.0" }
classifiers = [
  "License :: Other/Proprietary License",
]
dependencies = [
  "aiohttp",
  "pydantic>=2",
  "dbus-next",
  "playwright",
  "beautifulsoup4",
  "lxml",
  "trafilatura",
  "psutil",
  "watchdog",
  "orjson",
  "tenacity",
  "typer",
  "rich",
  "httpx>=0.27",
  "PyYAML",
  "numpy",
  "pystray",
  "Pillow>=10.0.0"
]

[project.optional-dependencies]
test = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.4.0",
  "black>=24.3.0",
]
ui = [
  "matplotlib>=3.7.0",
  "rapidfuzz>=3.6.1",
  "markdown>=3.6",
  "tkinterweb>=3.23.8",
]
hf = [
  "torch>=2.1.0; platform_system != 'Windows' or platform_machine != 'ARM64'",
  "transformers>=4.41.0",
  "accelerate>=0.31.0",
  "sentencepiece>=0.1.99",
  "protobuf>=3.20.0",  # Required for tokenizers
  "peft>=0.11.1",
  "datasets>=2.14.0",
  "huggingface_hub>=0.19.0",
  "hf-xet>=1.0.0",  # Xet Storage for faster HuggingFace dataset downloads
  "Pillow>=10.0.0",
  "tqdm>=4.65.0",
  "safetensors>=0.3.1",
  "deepspeed>=0.14.0; platform_system != 'Windows'",  # Windows: install manually from prebuilt wheel
  "mpi4py>=4.0.0; platform_system != 'Windows'",  # Required for DeepSpeed distributed training (Linux only)
  "bitsandbytes>=0.43.0",
  "flash-attn>=2.3.0; platform_system != 'Windows'",  # Windows requires manual install due to path limits
  "lm-eval>=0.4.0",  # For model evaluation benchmarks
]

[project.scripts]
aios = "aios.cli.aios:app"

[project.entry-points."lm_eval.models"]
aios = "aios.core.evaluation.aios_lm_eval_adapter"

[tool.setuptools]
packages = { find = { where = ["src"] } }

[tool.setuptools.package-dir]
"" = "src"
